requirements.txt
Flask==2.2.5
Pillow==9.4.0
google-generativeai==0.3.1
python-dotenv==1.0.0
